requirements.txt
# Core dependencies
python-dotenv==1.0.0
openai==1.58.1

# PDF and Image processing
PyMuPDF==1.23.26
pdf2image==1.17.0
Pillow==10.2.0

# OCR
pytesseract==0.3.10

# Utilities
tqdm==4.66.1
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
